Move cdx utils to separate file
